requirements.txt
botometer~=1.6
pandas~=1.1.2
numpy~=1.18.3
click~=7.1.2
python-dotenv~=0.14.0